requirement.txt
numpy
faiss-cpu
sentence-transformers
fastapi
uvicorn
sqlalchemy 
pymysql